Code_using_React/fastAPI/requirements.txt
uvicorn
fastapi[all]
python-dotenv
numpy
scikit-learn
pydantic
langchain
langchain-google-genai
langchain-groq
azure-devops
msrest